% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{classString}
\alias{classString}
\title{S3 class as a single string}
\usage{
classString(opts, collapse = "_")
}
\arguments{
\item{opts}{Any object, in particular an Opts object.}

\item{collapse}{An optional character string to separate the class vector
entries.}
}
\value{
A single string. The concatenation of the S3 class elements.
}
\description{
S3 class as a single string
}
