% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigOpts-package.R
\docType{package}
\name{ConfigOpts-package}
\alias{ConfigOpts}
\alias{ConfigOpts-package}
\title{ConfigOpts: S3 Class for Serializable Options}
\description{
Defines an S3 class Opts. It stores configuratio options that can be used as arguemnts for function calls.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
