% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{writeOpts}
\alias{writeOpts}
\title{Write an Opts object as a json file.}
\usage{
writeOpts(opts, file = NULL, dir = NULL, addMetaInfo = TRUE)
}
\arguments{
\item{opts}{An object of S3 class Opts.}

\item{file}{A path to the output file or \code{NULL}.}

\item{dir}{\code{NULL} or a path to the output directory. In the latter case, the
output file name will be created automatically from the class.}

\item{addMetaInfo}{Should a time stamp and package versions be added before
writing the object?}
}
\description{
Write an Opts object as a json file.
}
