% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{writeOpts}
\alias{writeOpts}
\title{Write an Opts object as a json file.}
\usage{
writeOpts(
  opts,
  file = NULL,
  dir = NULL,
  addMetaInfo = TRUE,
  digits = NA,
  warn = TRUE,
  validate = TRUE
)
}
\arguments{
\item{opts}{An object of S3 class Opts.}

\item{file}{A path to the output file or \code{NULL}.}

\item{dir}{\code{NULL} or a path to the output directory. In the latter case, the
output file name will be created automatically from the class.}

\item{addMetaInfo}{Should a time stamp and package versions be added before
writing the object?}

\item{digits}{max number of decimal digits to write for numeric values. Use
\code{I()} to specify significant digits. Use \code{NA} for max precision.}

\item{warn}{Should warnings be thrown if, e.g., a file will be overwritten?}

\item{validate}{Should the opts object be validated? Throws an error for invalid objects.}
}
\description{
Write an Opts object as a json file.
}
